client/games/secret_paint_game/web_game_runner/pyproject.toml
[project]
name = "web-game-runner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pywebview>=6.1",
]
